+++ scripts/build-styles.mjs
@@ -71,6 +71,7 @@ await writeFile(
     '@import "../../src/components/skeleton/skeleton.css";',
     '@import "../../src/components/progress/progress.css";',
     '@import "../../src/components/progress-circle/progress-circle.css";',
+    '@import "../../src/components/toast/toast.css";',
     '@import "../../src/components/collapsible/collapsible.css";',
     '@import "../../src/components/accordion/accordion.css";',
     '@import "../../src/components/input/input.css";',